# Optimize the allocation and speed of ActivitySet/GetCustomProperty