@macrogreg 2020-09-03
Thanks for doing this, @tarekgh .
This does not go quite as far as previously suggested in https://github.com/dotnet/runtime/pull/39690/ ,  but it goes a good step in that direction. Glad to see this improvement!