@CodeBlanch 2020-09-04
Thanks for doing this @tarekgh. Interesting results! Looks like Dictionary<,> uses less memory than ConcurrentDictionary<,>, that seems reasonable. The speed-up, that's surprising. Explicitly taking a lock is faster than the lockless algorithm? 🤯