@tarekgh 2020-09-04
> that's surprising. Explicitly taking a lock is faster than the lockless algorithm?

Usually the lockless will be faster in the scenarios when more than one thread competing on executing the atomic block. In Activity this is not regular case and the lock will be cheap especially we are using the internal dictionary instance as the lock.