@noahfalk 2020-09-05
> Explicitly taking a lock is faster than the lockless algorithm? 🤯

ConcurrentDictionary isn't lockless, the locks are just managed for you : ) https://source.dot.net/#System.Collections.Concurrent/System/Collections/Concurrent/ConcurrentDictionary.cs,916

I'd guess that if many threads reading/writing simultaneously then ConcurrentDictionary's multiple lock strategy is faster, but if the common case is only 1 reader/writer then the multi-lock overhead doesn't payoff. Activity is in the latter category. Fwiw I am a little surprised the BCL team didn't implement it lock-free but there is probably a good reason for their choice I am unaware of.