@tarekgh 2020-09-05
/backport to release/5.0-rc2