@stephentoub 2020-09-05
> ConcurrentDictionary isn't lockless, the locks are just managed for you : )

It uses striped locking for writes, but reads are lock-free. I expect if you were to measure the throughput of this new version just repeatedly reading custom properties (rather than measuring also writing them), you'd find the nee version is slower (with less memory).