@noahfalk 2020-09-05
> I expect if you were to measure the throughput of this new version just repeatedly reading custom properties

Thanks Stephen! I'd expect the typical usage to be a write:read ratio that relatively even, 1:1 or 1:2. I would consider cases that read many times to be our non-perf conscious consumers in this particular scenario so optimizing for fewer reads and lower memory usage makes sense.

For my own curiousity, do you know what the tradeoffs were between striped locks and a lock-free write? I searched around in the source a little bit but I couldn't find any documented rationale on the implementation choice.