@stephentoub 2020-09-05
> For my own curiousity, do you know what the tradeoffs were between striped locks and a lock-free write?

How would you implement the lock-free write?