@noahfalk 2020-09-05
> How would you implement the lock-free write?

I have never gone into the details but my limited understanding suggested implementations did exist if we wanted them? http://www.cse.chalmers.se/~tsigas/papers/Lock-Free_Dictionary.pdf
Whether the implementations were any good and/or applicable to particular requirements we had I have no idea. I was hoping someone else might have already done the research : )