@tarekgh 2020-09-05
https://github.com/VSadov/NonBlocking