@danmosemsft 2020-09-05
That’s interesting I wonder what advantages ConcurrentDictionary has if any over @vsadov one above.