@am11 2020-09-05
Thanks for sharing that link. Slides from the YouTube video in [readme](https://github.com/VSadov/NonBlocking#implementation-notes) are available here https://web.stanford.edu/class/ee380/Abstracts/070221_LockFreeHash.pdf.