@stephentoub 2020-09-05
> implementations did exist if we wanted them?

If we want to overhaul the entire implementation, sure. I'm not aware of a good way to implement lock-free writes with this implementation.

@vsadov and I discussed his implementation several years back. I have zero attachment to the current implementation; my main concern is ensuring that semantics and correctness are maintained while not regressing perf, especially for reads. If a new implementation can maintain all that and improve throughput and reduce memory allocation/consumption, great. I previously rewrote ConcurrentQueue for .NET Core 2.0 following similar principles.