@noahfalk 2020-09-09
Thanks all! From what I see the only downside yet presented for the non-blocking option is that it would need some work to make it the new default + inherent risk that a new implementation hasn't been validated in a shipping product. To me that sounds quite promising and would just need to be prioritized against other BCL perf improvements we'd spend our time on. The scenario in this issue is probably far too narrow to be the sole justification, but across .NET broadly its not hard too imagine that many scenarios would all benefit a bit from a faster concurrent dictionary. If I spot other instances that would benefit I'll give a heads up.